% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{create_api}
\alias{create_api}
\title{Creates an Api resource}
\usage{
create_api(
  ApiKeySelectionExpression = "string",
  Description = "string",
  DisableSchemaValidation = TRUE|FALSE,
  Name = "string",
  ProtocolType = "WEBSOCKET",
  RouteSelectionExpression = "string",
  Version = "string"
)
}
\arguments{
\item{ApiKeySelectionExpression}{An API key selection expression. See \href{https://docs.aws.amazon.com/apigateway/latest/developerguide/apigateway-websocket-api-selection-expressions.html#apigateway-websocket-api-apikey-selection-expressions}{API Key Selection Expressions}.}

\item{Description}{The description of the API.}

\item{DisableSchemaValidation}{Avoid validating models when creating a deployment.}

\item{Name}{[required] The name of the API.}

\item{ProtocolType}{[required] The API protocol: HTTP or WEBSOCKET.}

\item{RouteSelectionExpression}{[required] The route selection expression for the API.}

\item{Version}{A version identifier for the API.}
}
\description{
Creates an Api resource.
}
\examples{

}
