% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{create_api_mapping}
\alias{create_api_mapping}
\title{Creates an API mapping}
\usage{
create_api_mapping(ApiId, ApiMappingKey = NULL, DomainName, Stage)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ApiMappingKey}{}

\item{DomainName}{[required] The domain name.}

\item{Stage}{[required] The API stage.}
}
\description{
Creates an API mapping.
}
\section{Accepted Parameters}{
\preformatted{create_api_mapping(
  ApiId = "string",
  ApiMappingKey = "string",
  DomainName = "string",
  Stage = "string"
)
}
}

