% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{create_route_response}
\alias{create_route_response}
\title{Creates a RouteResponse for a Route}
\usage{
create_route_response(ApiId, ModelSelectionExpression = NULL,
  ResponseModels = NULL, ResponseParameters = NULL, RouteId,
  RouteResponseKey)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelSelectionExpression}{The model selection expression for the route response.}

\item{ResponseModels}{The response models for the route response.}

\item{ResponseParameters}{The route response parameters.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseKey}{[required] The route response key.}
}
\description{
Creates a RouteResponse for a Route.
}
\section{Accepted Parameters}{
\preformatted{create_route_response(
  ApiId = "string",
  ModelSelectionExpression = "string",
  ResponseModels = list(
    "string"
  ),
  ResponseParameters = list(
    list(
      Required = TRUE|FALSE
    )
  ),
  RouteId = "string",
  RouteResponseKey = "string"
)
}
}

