% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_api}
\alias{delete_api}
\title{Deletes an Api resource}
\usage{
delete_api(ApiId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}
}
\description{
Deletes an Api resource.
}
\section{Accepted Parameters}{
\preformatted{delete_api(
  ApiId = "string"
)
}
}

