% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_api_mapping}
\alias{delete_api_mapping}
\title{Deletes an API mapping}
\usage{
delete_api_mapping(
  ApiId = "string",
  ApiMappingId = "string",
  DomainName = "string"
)
}
\arguments{
\item{ApiId}{[required] The identifier of the API.}

\item{ApiMappingId}{[required] The API mapping identifier.}

\item{DomainName}{[required] The domain name.}
}
\description{
Deletes an API mapping.
}
\examples{

}
