% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_authorizer}
\alias{delete_authorizer}
\title{Deletes an Authorizer}
\usage{
delete_authorizer(ApiId, AuthorizerId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{AuthorizerId}{[required] The authorizer identifier.}
}
\description{
Deletes an Authorizer.
}
\section{Accepted Parameters}{
\preformatted{delete_authorizer(
  ApiId = "string",
  AuthorizerId = "string"
)
}
}

