% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_deployment}
\alias{delete_deployment}
\title{Deletes a Deployment}
\usage{
delete_deployment(ApiId, DeploymentId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{DeploymentId}{[required] The deployment ID.}
}
\description{
Deletes a Deployment.
}
\section{Accepted Parameters}{
\preformatted{delete_deployment(
  ApiId = "string",
  DeploymentId = "string"
)
}
}

