% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_integration}
\alias{delete_integration}
\title{Deletes an Integration}
\usage{
delete_integration(
  ApiId = "string",
  IntegrationId = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}
}
\description{
Deletes an Integration.
}
\examples{

}
