% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_model}
\alias{delete_model}
\title{Deletes a Model}
\usage{
delete_model(ApiId, ModelId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelId}{[required] The model ID.}
}
\description{
Deletes a Model.
}
\section{Accepted Parameters}{
\preformatted{delete_model(
  ApiId = "string",
  ModelId = "string"
)
}
}

