% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_route}
\alias{delete_route}
\title{Deletes a Route}
\usage{
delete_route(
  ApiId = "string",
  RouteId = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RouteId}{[required] The route ID.}
}
\description{
Deletes a Route.
}
\examples{

}
