% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_api_mappings}
\alias{get_api_mappings}
\title{The API mappings}
\usage{
get_api_mappings(DomainName, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{DomainName}{[required] The domain name.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last element of the collection.}
}
\description{
The API mappings.
}
\section{Accepted Parameters}{
\preformatted{get_api_mappings(
  DomainName = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

