% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_apis}
\alias{get_apis}
\title{Gets a collection of Api resources}
\usage{
get_apis(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last element of the collection.}
}
\description{
Gets a collection of Api resources.
}
\section{Accepted Parameters}{
\preformatted{get_apis(
  MaxResults = "string",
  NextToken = "string"
)
}
}

