% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_deployment}
\alias{get_deployment}
\title{Gets a Deployment}
\usage{
get_deployment(ApiId, DeploymentId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{DeploymentId}{[required] The deployment ID.}
}
\description{
Gets a Deployment.
}
\section{Accepted Parameters}{
\preformatted{get_deployment(
  ApiId = "string",
  DeploymentId = "string"
)
}
}

