% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_domain_names}
\alias{get_domain_names}
\title{Gets the domain names for an AWS account}
\usage{
get_domain_names(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last element of the collection.}
}
\description{
Gets the domain names for an AWS account.
}
\section{Accepted Parameters}{
\preformatted{get_domain_names(
  MaxResults = "string",
  NextToken = "string"
)
}
}

