% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_integration}
\alias{get_integration}
\title{Gets an Integration}
\usage{
get_integration(
  ApiId = "string",
  IntegrationId = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}
}
\description{
Gets an Integration.
}
\examples{

}
