% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_integration_response}
\alias{get_integration_response}
\title{Gets an IntegrationResponses}
\usage{
get_integration_response(
  ApiId = "string",
  IntegrationId = "string",
  IntegrationResponseId = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}

\item{IntegrationResponseId}{[required] The integration response ID.}
}
\description{
Gets an IntegrationResponses.
}
\examples{

}
