% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_integration_responses}
\alias{get_integration_responses}
\title{Gets the IntegrationResponses for an Integration}
\usage{
get_integration_responses(ApiId, IntegrationId, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last element of the collection.}
}
\description{
Gets the IntegrationResponses for an Integration.
}
\section{Accepted Parameters}{
\preformatted{get_integration_responses(
  ApiId = "string",
  IntegrationId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

