% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_model_template}
\alias{get_model_template}
\title{Gets a model template}
\usage{
get_model_template(
  ApiId = "string",
  ModelId = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelId}{[required] The model ID.}
}
\description{
Gets a model template.
}
\examples{

}
