% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_route}
\alias{get_route}
\title{Gets a Route}
\usage{
get_route(ApiId, RouteId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RouteId}{[required] The route ID.}
}
\description{
Gets a Route.
}
\section{Accepted Parameters}{
\preformatted{get_route(
  ApiId = "string",
  RouteId = "string"
)
}
}

