% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_route_response}
\alias{get_route_response}
\title{Gets a RouteResponse}
\usage{
get_route_response(ApiId, RouteId, RouteResponseId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseId}{[required] The route response ID.}
}
\description{
Gets a RouteResponse.
}
\section{Accepted Parameters}{
\preformatted{get_route_response(
  ApiId = "string",
  RouteId = "string",
  RouteResponseId = "string"
)
}
}

