% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_route_responses}
\alias{get_route_responses}
\title{Gets the RouteResponses for a Route}
\usage{
get_route_responses(ApiId, MaxResults = NULL, NextToken = NULL,
  RouteId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last element of the collection.}

\item{RouteId}{[required] The route ID.}
}
\description{
Gets the RouteResponses for a Route.
}
\section{Accepted Parameters}{
\preformatted{get_route_responses(
  ApiId = "string",
  MaxResults = "string",
  NextToken = "string",
  RouteId = "string"
)
}
}

