% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_stage}
\alias{get_stage}
\title{Gets a Stage}
\usage{
get_stage(ApiId, StageName)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{StageName}{[required] The stage name.}
}
\description{
Gets a Stage.
}
\section{Accepted Parameters}{
\preformatted{get_stage(
  ApiId = "string",
  StageName = "string"
)
}
}

