% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_stages}
\alias{get_stages}
\title{Gets the Stages for an API}
\usage{
get_stages(
  ApiId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last element of the collection.}
}
\description{
Gets the Stages for an API.
}
\examples{

}
