% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_package.R
\docType{package}
\name{paws.apigatewayv2-package}
\alias{paws.apigatewayv2}
\alias{paws.apigatewayv2-package}
\title{paws.apigatewayv2: AmazonApiGatewayV2}
\description{
Amazon API Gateway V2
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
