% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{update_api_mapping}
\alias{update_api_mapping}
\title{The API mapping}
\usage{
update_api_mapping(
  ApiId = "string",
  ApiMappingId = "string",
  ApiMappingKey = "string",
  DomainName = "string",
  Stage = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ApiMappingId}{[required] The API mapping identifier.}

\item{ApiMappingKey}{The API mapping key.}

\item{DomainName}{[required] The domain name.}

\item{Stage}{The API stage.}
}
\description{
The API mapping.
}
\examples{

}
