% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{update_deployment}
\alias{update_deployment}
\title{Updates a Deployment}
\usage{
update_deployment(
  ApiId = "string",
  DeploymentId = "string",
  Description = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{DeploymentId}{[required] The deployment ID.}

\item{Description}{The description for the deployment resource.}
}
\description{
Updates a Deployment.
}
\examples{

}
