% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{update_domain_name}
\alias{update_domain_name}
\title{Updates a domain name}
\usage{
update_domain_name(DomainName, DomainNameConfigurations = NULL)
}
\arguments{
\item{DomainName}{[required] The domain name.}

\item{DomainNameConfigurations}{The domain name configurations.}
}
\description{
Updates a domain name.
}
\section{Accepted Parameters}{
\preformatted{update_domain_name(
  DomainName = "string",
  DomainNameConfigurations = list(
    list(
      ApiGatewayDomainName = "string",
      CertificateArn = "string",
      CertificateName = "string",
      CertificateUploadDate = as.POSIXct("2015-01-01"),
      EndpointType = "REGIONAL"|"EDGE",
      HostedZoneId = "string"
    )
  )
)
}
}

