% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{update_model}
\alias{update_model}
\title{Updates a Model}
\usage{
update_model(
  ApiId = "string",
  ContentType = "string",
  Description = "string",
  ModelId = "string",
  Name = "string",
  Schema = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ContentType}{The content-type for the model, for example, "application/json".}

\item{Description}{The description of the model.}

\item{ModelId}{[required] The model ID.}

\item{Name}{The name of the model.}

\item{Schema}{The schema for the model. For application/json models, this should be JSON schema draft 4 model.}
}
\description{
Updates a Model.
}
\examples{

}
