% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{update_route_response}
\alias{update_route_response}
\title{Updates a RouteResponse}
\usage{
update_route_response(ApiId, ModelSelectionExpression = NULL,
  ResponseModels = NULL, ResponseParameters = NULL, RouteId,
  RouteResponseId, RouteResponseKey = NULL)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelSelectionExpression}{The model selection expression for the route response.}

\item{ResponseModels}{The response models for the route response.}

\item{ResponseParameters}{The route response parameters.}

\item{RouteId}{[required] The route ID.}

\item{RouteResponseId}{[required] The route response ID.}

\item{RouteResponseKey}{The route response key.}
}
\description{
Updates a RouteResponse.
}
\section{Accepted Parameters}{
\preformatted{update_route_response(
  ApiId = "string",
  ModelSelectionExpression = "string",
  ResponseModels = list(
    "string"
  ),
  ResponseParameters = list(
    list(
      Required = TRUE|FALSE
    )
  ),
  RouteId = "string",
  RouteResponseId = "string",
  RouteResponseKey = "string"
)
}
}

