% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{create_deployment}
\alias{create_deployment}
\title{Creates a Deployment for an API}
\usage{
create_deployment(ApiId, Description = NULL, StageName = NULL)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{Description}{The description for the deployment resource.}

\item{StageName}{The name of the Stage resource for the Deployment resource to create.}
}
\description{
Creates a Deployment for an API.
}
\section{Accepted Parameters}{
\preformatted{create_deployment(
  ApiId = "string",
  Description = "string",
  StageName = "string"
)
}
}

