% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_domain_name}
\alias{delete_domain_name}
\title{Deletes a domain name}
\usage{
delete_domain_name(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain name.}
}
\description{
Deletes a domain name.
}
\section{Accepted Parameters}{
\preformatted{delete_domain_name(
  DomainName = "string"
)
}
}

