% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{delete_integration_response}
\alias{delete_integration_response}
\title{Deletes an IntegrationResponses}
\usage{
delete_integration_response(ApiId, IntegrationId, IntegrationResponseId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{IntegrationId}{[required] The integration ID.}

\item{IntegrationResponseId}{[required] The integration response ID.}
}
\description{
Deletes an IntegrationResponses.
}
\section{Accepted Parameters}{
\preformatted{delete_integration_response(
  ApiId = "string",
  IntegrationId = "string",
  IntegrationResponseId = "string"
)
}
}

