% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_api}
\alias{get_api}
\title{Gets an Api resource}
\usage{
get_api(
  ApiId = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}
}
\description{
Gets an Api resource.
}
\examples{

}
