% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_api_mapping}
\alias{get_api_mapping}
\title{The API mapping}
\usage{
get_api_mapping(
  ApiId = "string",
  ApiMappingId = "string",
  DomainName = "string"
)
}
\arguments{
\item{ApiId}{[required] The identifier of the API.}

\item{ApiMappingId}{[required] The API mapping identifier.}

\item{DomainName}{[required] The domain name.}
}
\description{
The API mapping.
}
\examples{

}
