% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_authorizer}
\alias{get_authorizer}
\title{Gets an Authorizer}
\usage{
get_authorizer(
  ApiId = "string",
  AuthorizerId = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{AuthorizerId}{[required] The authorizer identifier.}
}
\description{
Gets an Authorizer.
}
\examples{

}
