% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_deployments}
\alias{get_deployments}
\title{Gets the Deployments for an API}
\usage{
get_deployments(ApiId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{MaxResults}{The maximum number of elements to be returned for this resource.}

\item{NextToken}{The next page of elements from this collection. Not valid for the last element of the collection.}
}
\description{
Gets the Deployments for an API.
}
\section{Accepted Parameters}{
\preformatted{get_deployments(
  ApiId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

