% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_domain_name}
\alias{get_domain_name}
\title{Gets a domain name}
\usage{
get_domain_name(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain name.}
}
\description{
Gets a domain name.
}
\section{Accepted Parameters}{
\preformatted{get_domain_name(
  DomainName = "string"
)
}
}

