% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewayv2_operations.R
\name{get_model}
\alias{get_model}
\title{Gets a Model}
\usage{
get_model(
  ApiId = "string",
  ModelId = "string"
)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelId}{[required] The model ID.}
}
\description{
Gets a Model.
}
\examples{

}
